Moved modules out of extracters and into pipeline hierarchy.
